+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added functions to AutoreleasingCache
     - contains
     - clear
+    - setNewExpiration
+- Added exception type for when a path in AutoreleasingCache does not exist yet was attempted to be read. (not breaking anything yet!)
 
 ## 1.1.3